refactor(engine): remove the unwired MIMD adaptive-resizing dead code
The concurrent-delta reorder buffer carried an opt-in adaptive
capacity-scaling policy (AdaptiveCapacityPolicy / AdaptiveState /
ReorderStats) that grew and shrank the ring at runtime under a
grow/shrink utilization controller. It was never wired into any
production path: the only callers of ReorderBuffer::with_adaptive_policy
and SpillableReorderBuffer::with_adaptive_policy were unit tests, so the
adaptive branch was dead everywhere the transfer engine actually runs.

Fixed-capacity ring sizing is the settled direction, so the runtime
resizer has no future and is deleted rather than revived:

- delete crates/engine/src/concurrent_delta/adaptive.rs
- drop the adaptive field, with_adaptive_policy, try_adaptive_preinsert_grow,
  maybe_adapt_capacity, and stats() from ReorderBuffer; simplify grow()
  to its doubling fallback
- drop with_adaptive_policy and reorder_stats from SpillableReorderBuffer
- remove the now-dead re-exports and adaptive unit tests, keeping the
  fixed-capacity in_flight_window / metrics observability tests (renamed
  the test module to observability.rs to match)

The fixed-capacity path (ReorderBuffer::new / passthrough) is untouched,
so delivery ordering and observable behaviour are identical. Build and
clippy are clean for the engine crate; the retained reorder, passthrough,
and observability tests pass.
