+++ crates/engine/src/concurrent_delta/reorder/tests.rs
@@ -832,185 +832,6 @@ mod core_tests {
     }
 }
 
-mod adaptive_tests {
-    use super::super::*;
-
-    /// Default-constructed buffers must remain unaffected by adaptive logic.
-    #[test]
-    fn fixed_capacity_default_unchanged() {
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::new(4);
-        for i in 0..4 {
-            buf.insert(i, i).unwrap();
-        }
-        assert_eq!(buf.capacity(), 4);
-        let stats = buf.stats();
-        assert_eq!(stats.grow_events, 0);
-        assert_eq!(stats.shrink_events, 0);
-        assert_eq!(stats.capacity, 4);
-        // Capacity-exceeded behaviour preserved.
-        assert_eq!(buf.insert(4, 4), Err(CapacityExceeded));
-    }
-
-    #[test]
-    fn adaptive_buffer_starts_at_min_capacity() {
-        let policy = AdaptiveCapacityPolicy::new(4, 32, 2.0);
-        let buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-        assert_eq!(buf.capacity(), 4);
-        assert_eq!(buf.stats().grow_events, 0);
-    }
-
-    /// Inserting beyond `min` capacity grows the ring (without losing items)
-    /// up to but never beyond `max`.
-    #[test]
-    fn grows_under_load() {
-        let policy = AdaptiveCapacityPolicy::with_window(2, 16, 2.0, 64);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        // Build a wide gap (insert seq 0..7 with a hole at 0) to force growth.
-        for seq in 1..8 {
-            // seq 1, 2 fit (capacity 2 -> grows). Subsequent inserts continue
-            // to grow up to max as the gap window widens.
-            buf.insert(seq, seq).unwrap();
-        }
-        let stats = buf.stats();
-        assert!(stats.grow_events > 0, "buffer never grew under load");
-        assert!(
-            buf.capacity() >= 8,
-            "capacity {} < required gap window 8",
-            buf.capacity()
-        );
-        assert!(
-            buf.capacity() <= 16,
-            "capacity {} exceeded max",
-            buf.capacity()
-        );
-
-        // Fill the head and confirm ordered drain still works after growth.
-        buf.insert(0, 0).unwrap();
-        let drained: Vec<u64> = buf.drain_ready().collect();
-        assert_eq!(drained, (0..8).collect::<Vec<_>>());
-    }
-
-    /// The grow path must never breach `policy.max`.
-    #[test]
-    fn never_exceeds_max() {
-        let policy = AdaptiveCapacityPolicy::with_window(2, 8, 2.0, 64);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        // Try to drive a gap of 12 - capacity should cap at 8.
-        for seq in 1..8 {
-            buf.insert(seq, seq).unwrap();
-        }
-        // Once at max, inserts beyond capacity must error.
-        let err = buf.insert(8, 8);
-        assert_eq!(err, Err(CapacityExceeded));
-        assert!(buf.capacity() <= 8);
-    }
-
-    /// After a sustained idle window, capacity should shrink back toward `min`.
-    #[test]
-    fn shrinks_when_idle() {
-        let policy = AdaptiveCapacityPolicy::with_window(2, 32, 2.0, 4);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        // Force a grow: build a gap so capacity expands.
-        for seq in 1..6 {
-            buf.insert(seq, seq).unwrap();
-        }
-        let grown = buf.capacity();
-        assert!(grown >= 6, "expected grown capacity, got {grown}");
-        let grow_events = buf.stats().grow_events;
-        assert!(grow_events >= 1);
-
-        // Drain everything to clear the gap.
-        buf.insert(0, 0).unwrap();
-        let _: Vec<_> = buf.drain_ready().collect();
-        assert!(buf.is_empty());
-
-        // Submit single-item inserts that immediately drain - low utilization.
-        for seq in 6..30 {
-            buf.insert(seq, seq).unwrap();
-            let _ = buf.next_in_order();
-        }
-        let stats = buf.stats();
-        assert!(stats.shrink_events >= 1, "buffer never shrank when idle");
-        assert!(buf.capacity() < grown, "capacity did not decrease");
-    }
-
-    /// Shrinks must clamp at `policy.min` no matter how long the idle window.
-    #[test]
-    fn never_drops_below_min() {
-        let min = 4usize;
-        let policy = AdaptiveCapacityPolicy::with_window(min, 32, 2.0, 4);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        // Drive growth then quiesce.
-        for seq in 1..10 {
-            buf.insert(seq, seq).unwrap();
-        }
-        buf.insert(0, 0).unwrap();
-        let _: Vec<_> = buf.drain_ready().collect();
-
-        for seq in 10..200 {
-            buf.insert(seq, seq).unwrap();
-            let _ = buf.next_in_order();
-        }
-        assert!(
-            buf.capacity() >= min,
-            "capacity {} dropped below min {min}",
-            buf.capacity()
-        );
-        // Min is the hard floor.
-        assert_eq!(buf.capacity(), min);
-    }
-
-    /// Stats reflect both grow and shrink events accurately.
-    #[test]
-    fn stats_track_both_events() {
-        let policy = AdaptiveCapacityPolicy::with_window(2, 16, 2.0, 4);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        // Trigger at least one grow.
-        for seq in 1..6 {
-            buf.insert(seq, seq).unwrap();
-        }
-        assert!(buf.stats().grow_events >= 1);
-
-        // Drain and idle to trigger at least one shrink.
-        buf.insert(0, 0).unwrap();
-        let _: Vec<_> = buf.drain_ready().collect();
-        for seq in 6..40 {
-            buf.insert(seq, seq).unwrap();
-            let _ = buf.next_in_order();
-        }
-        let stats = buf.stats();
-        assert!(stats.grow_events >= 1);
-        assert!(stats.shrink_events >= 1);
-        assert_eq!(stats.capacity, buf.capacity());
-    }
-
-    /// Ordering is preserved across grow / shrink transitions.
-    #[test]
-    fn ordering_preserved_through_resize() {
-        let policy = AdaptiveCapacityPolicy::with_window(2, 64, 2.0, 8);
-        let mut buf: ReorderBuffer<u64> = ReorderBuffer::with_adaptive_policy(policy);
-
-        let n: u64 = 200;
-        // Insert in reversed bursts of 8 to force out-of-order delivery and
-        // exercise both grow and shrink paths.
-        let mut collected: Vec<u64> = Vec::with_capacity(n as usize);
-        for burst_start in (0..n).step_by(8) {
-            let end = (burst_start + 8).min(n);
-            for seq in (burst_start..end).rev() {
-                buf.insert(seq, seq).unwrap();
-            }
-            collected.extend(buf.drain_ready());
-        }
-        let expected: Vec<u64> = (0..n).collect();
-        assert_eq!(collected, expected);
-    }
-}
-
 mod passthrough_tests {
     use super::super::*;
 
@@ -1109,15 +930,6 @@ mod passthrough_tests {
     fn passthrough_capacity_is_zero() {
         let buf: ReorderBuffer<u64> = ReorderBuffer::passthrough();
         assert_eq!(buf.capacity(), 0);
-    }
-
-    #[test]
-    fn passthrough_stats_show_no_adaptive_events() {
-        let buf: ReorderBuffer<u64> = ReorderBuffer::passthrough();
-        let stats = buf.stats();
-        assert_eq!(stats.grow_events, 0);
-        assert_eq!(stats.shrink_events, 0);
-        assert_eq!(stats.capacity, 0);
     }
 
     #[test]
